+++ misc/pandas_utils.py
@@ -7,8 +7,15 @@
 #   hold various types of records.
 ###
 
+import gzip
+import os
+import shutil
+
 import numpy as np
 import pandas as pd
+
+import openpyxl
+import fastparquet
 
 import logging
 logger = logging.getLogger(__name__)
@@ -33,8 +40,6 @@ def dict_to_dataframe(dic, key_name='key', value_name='value'):
     df: pd.DataFrame
         a data frame in which each row corresponds to one entry in dic
     """
-    import pandas as pd
-
     df = pd.Series(dic, name=value_name)
     df.index.name = key_name
     df = df.reset_index()
@@ -60,6 +65,7 @@ def dataframe_to_dict(df, key_field, value_field):
 
 excel_extensions = ('xls', 'xlsx')
 hdf5_extensions = ('hdf', 'hdf5', 'h5', 'he5')
+parquet_extensions = ('parq', )
 
 def _guess_df_filetype(filename):
     """ This function attempts to guess the filetype given a filename. It is
@@ -68,6 +74,7 @@ def _guess_df_filetype(filename):
 
             excel: xls, xlsx
             hdf5: hdf, hdf5, h5, he5
+            parquet: parq
             csv: all other extensions
 
         Additionally, if filename is a pd.ExcelWriter object, then the guessed
@@ -83,8 +90,6 @@ def _guess_df_filetype(filename):
         Imports:
             pandas
     """
-    import pandas as pd
-
     msg = "Attempting to guess the extension. Filename: {}".format(filename)
     logger.debug(msg)
 
@@ -94,6 +99,8 @@ def _guess_df_filetype(filename):
         filetype = 'excel'
     elif filename.endswith(hdf5_extensions):
         filetype= 'hdf5'
+    elif filename.endswith(parquet_extensions):
+        filetype= 'parquet'
     else:
         filetype = 'csv'
 
@@ -110,6 +117,7 @@ def read_df(filename, filetype='AUTO', sheet=None, **kwargs):
 
             excel: xls, xlsx
             hdf5: hdf, hdf5, h5, he5
+            parquet: parq
             csv: all other extensions
 
         N.B. In principle, matlab data files are hdf5, so this function should
@@ -136,11 +144,7 @@ def read_df(filename, filetype='AUTO', sheet=None, **kwargs):
             ValueError: if the filetype is not 'AUTO' or one of the values
                 mentioned above ('excel', 'hdf5', 'csv')
 
-        Imports:
-            pandas
     """
-    import pandas as pd
-
     # first, see if we want to guess the filetype
     if filetype == 'AUTO':
         filetype = _guess_df_filetype(filename)
@@ -152,6 +156,9 @@ def read_df(filename, filetype='AUTO', sheet=None, **kwargs):
         df = pd.read_excel(filename, sheetname=sheet, **kwargs)
     elif filetype == 'hdf5':
         df = pd.read_hdf(filename, key=sheet, **kwargs)
+    elif filetype == "parquet":
+        pf = fastparquet.ParquetFile(filename, **kwargs)
+        df = pf.to_pandas()
     else:
         msg = "Could not read dataframe. Invalid filetype: {}".format(filetype)
         raise ValueError(msg)
@@ -167,6 +174,7 @@ def write_df(df, out, create_path=False, filetype='AUTO', sheet='Sheet_1',
 
             excel: xls, xlsx
             hdf5: hdf, hdf5, h5, he5
+            parquet: parq
             csv: all other extensions (e.g., "gz" or "bed")
 
         Additionally, the filetype can be specified as 'excel_writer'. In this
@@ -213,8 +221,6 @@ def write_df(df, out, create_path=False, filetype='AUTO', sheet='Sheet_1',
         -------
         None, but the file is created
     """
-    import gzip
-    import pandas as pd
     
     # first, see if we want to guess the filetype
     if filetype == 'AUTO':
@@ -247,6 +253,22 @@ def write_df(df, out, create_path=False, filetype='AUTO', sheet='Sheet_1',
 
     elif filetype == 'hdf5':
         df.to_hdf(out, sheet, **kwargs)
+
+    elif filetype == 'parquet':
+        if not do_not_compress:
+            kwargs['compression'] = 'GZIP'
+
+        # if a parquet "file" exists, delete it
+        if os.path.exists(out):
+            # it could be either a folder or a file
+            if os.path.isfile(out):
+                # delete file
+                os.remove(out)
+            else:
+                # delete directory
+                shutil.rmtree(out)
+        fastparquet.write(out, df, **kwargs)
+
     else:
         msg = ("Could not write the dataframe. Invalid filetype: {}".format(
             filetype))
@@ -276,9 +298,6 @@ def append_to_xlsx(df, xlsx, sheet='Sheet_1', **kwargs):
             pandas
             openpyxl
     """
-    import os
-    import pandas as pd
-    import openpyxl
 
     # check if the file already exists
     if os.path.exists(xlsx):