+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this repo will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/), 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.2.3] - 2017-07-27
+### Added
+- Helper to remove negative values from an np.array
+
+### Updated
+- Support for parquet files in `pandas_utils`
+
 ## [0.2.2] - 2017-07-24
 ### Added
 - A helper for controlling dask clusters